+++ tests/e2e/transaction/transactionSimulation.test.ts
@@ -127,12 +127,13 @@ describe("transaction simulation", () => {
       test("with script payload", async () => {
         const rawTxn = await aptos.generateTransaction({
           sender: singleSignerED25519SenderAccount.accountAddress.toString(),
-          feePayerAddress: feePayerAccount.accountAddress.toString(),
+          hasFeePayer: true,
           data: {
             bytecode: singleSignerScriptBytecode,
             functionArguments: [new U64(1), recieverAccounts[0].accountAddress],
           },
         });
+        rawTxn.feePayerAddress = feePayerAccount.accountAddress;
 
         const [response] = await aptos.simulateTransaction({
           signerPublicKey: singleSignerED25519SenderAccount.publicKey,
@@ -144,12 +145,14 @@ describe("transaction simulation", () => {
       test("with entry function payload", async () => {
         const rawTxn = await aptos.generateTransaction({
           sender: singleSignerED25519SenderAccount.accountAddress.toString(),
-          feePayerAddress: feePayerAccount.accountAddress.toString(),
+          hasFeePayer: true,
           data: {
             function: `${contractPublisherAccount.accountAddress.toString()}::transfer::transfer`,
             functionArguments: [new U64(1), recieverAccounts[0].accountAddress],
           },
         });
+        rawTxn.feePayerAddress = feePayerAccount.accountAddress;
+
         const [response] = await aptos.simulateTransaction({
           signerPublicKey: singleSignerED25519SenderAccount.publicKey,
           transaction: rawTxn,
@@ -160,13 +163,14 @@ describe("transaction simulation", () => {
       test("with multisig payload", async () => {
         const rawTxn = await aptos.generateTransaction({
           sender: singleSignerED25519SenderAccount.accountAddress.toString(),
-          feePayerAddress: feePayerAccount.accountAddress.toString(),
+          hasFeePayer: true,
           data: {
             multisigAddress: secondarySignerAccount.accountAddress,
             function: `${contractPublisherAccount.accountAddress.toString()}::transfer::transfer`,
             functionArguments: [new U64(1), recieverAccounts[0].accountAddress],
           },
         });
+        rawTxn.feePayerAddress = feePayerAccount.accountAddress;
 
         const [response] = await aptos.simulateTransaction({
           signerPublicKey: singleSignerED25519SenderAccount.publicKey,
@@ -179,7 +183,7 @@ describe("transaction simulation", () => {
         const rawTxn = await aptos.generateTransaction({
           sender: singleSignerED25519SenderAccount.accountAddress.toString(),
           secondarySignerAddresses: [secondarySignerAccount.accountAddress.toString()],
-          feePayerAddress: feePayerAccount.accountAddress.toString(),
+          hasFeePayer: true,
           data: {
             function: `${contractPublisherAccount.accountAddress.toString()}::transfer::two_by_two`,
             functionArguments: [
@@ -191,6 +195,7 @@ describe("transaction simulation", () => {
             ],
           },
         });
+        rawTxn.feePayerAddress = feePayerAccount.accountAddress;
 
         const [response] = await aptos.simulateTransaction({
           signerPublicKey: singleSignerED25519SenderAccount.publicKey,
@@ -305,12 +310,13 @@ describe("transaction simulation", () => {
       test("with script payload", async () => {
         const rawTxn = await aptos.generateTransaction({
           sender: singleSignerSecp256k1Account.accountAddress.toString(),
-          feePayerAddress: feePayerAccount.accountAddress.toString(),
+          hasFeePayer: true,
           data: {
             bytecode: singleSignerScriptBytecode,
             functionArguments: [new U64(1), recieverAccounts[0].accountAddress],
           },
         });
+        rawTxn.feePayerAddress = feePayerAccount.accountAddress;
 
         const [response] = await aptos.simulateTransaction({
           signerPublicKey: singleSignerSecp256k1Account.publicKey,
@@ -322,12 +328,14 @@ describe("transaction simulation", () => {
       test("with entry function payload", async () => {
         const rawTxn = await aptos.generateTransaction({
           sender: singleSignerSecp256k1Account.accountAddress.toString(),
-          feePayerAddress: feePayerAccount.accountAddress.toString(),
+          hasFeePayer: true,
           data: {
             function: `${contractPublisherAccount.accountAddress.toString()}::transfer::transfer`,
             functionArguments: [new U64(1), recieverAccounts[0].accountAddress],
           },
         });
+        rawTxn.feePayerAddress = feePayerAccount.accountAddress;
+
         const [response] = await aptos.simulateTransaction({
           signerPublicKey: singleSignerSecp256k1Account.publicKey,
           transaction: rawTxn,
@@ -338,13 +346,14 @@ describe("transaction simulation", () => {
       test("with multisig payload", async () => {
         const rawTxn = await aptos.generateTransaction({
           sender: singleSignerSecp256k1Account.accountAddress.toString(),
-          feePayerAddress: feePayerAccount.accountAddress.toString(),
+          hasFeePayer: true,
           data: {
             multisigAddress: secondarySignerAccount.accountAddress,
             function: `${contractPublisherAccount.accountAddress.toString()}::transfer::transfer`,
             functionArguments: [new U64(1), recieverAccounts[0].accountAddress],
           },
         });
+        rawTxn.feePayerAddress = feePayerAccount.accountAddress;
 
         const [response] = await aptos.simulateTransaction({
           signerPublicKey: singleSignerSecp256k1Account.publicKey,
@@ -357,7 +366,7 @@ describe("transaction simulation", () => {
         const rawTxn = await aptos.generateTransaction({
           sender: singleSignerSecp256k1Account.accountAddress.toString(),
           secondarySignerAddresses: [secondarySignerAccount.accountAddress.toString()],
-          feePayerAddress: feePayerAccount.accountAddress.toString(),
+          hasFeePayer: true,
           data: {
             function: `${contractPublisherAccount.accountAddress.toString()}::transfer::two_by_two`,
             functionArguments: [
@@ -369,6 +378,7 @@ describe("transaction simulation", () => {
             ],
           },
         });
+        rawTxn.feePayerAddress = feePayerAccount.accountAddress;
 
         const [response] = await aptos.simulateTransaction({
           signerPublicKey: singleSignerSecp256k1Account.publicKey,
@@ -483,12 +493,13 @@ describe("transaction simulation", () => {
       test("with script payload", async () => {
         const rawTxn = await aptos.generateTransaction({
           sender: legacyED25519SenderAccount.accountAddress.toString(),
-          feePayerAddress: feePayerAccount.accountAddress.toString(),
+          hasFeePayer: true,
           data: {
             bytecode: singleSignerScriptBytecode,
             functionArguments: [new U64(1), recieverAccounts[0].accountAddress],
           },
         });
+        rawTxn.feePayerAddress = feePayerAccount.accountAddress;
 
         const [response] = await aptos.simulateTransaction({
           signerPublicKey: legacyED25519SenderAccount.publicKey,
@@ -500,12 +511,14 @@ describe("transaction simulation", () => {
       test("with entry function payload", async () => {
         const rawTxn = await aptos.generateTransaction({
           sender: legacyED25519SenderAccount.accountAddress.toString(),
-          feePayerAddress: feePayerAccount.accountAddress.toString(),
+          hasFeePayer: true,
           data: {
             function: `${contractPublisherAccount.accountAddress.toString()}::transfer::transfer`,
             functionArguments: [new U64(1), recieverAccounts[0].accountAddress],
           },
         });
+        rawTxn.feePayerAddress = feePayerAccount.accountAddress;
+
         const [response] = await aptos.simulateTransaction({
           signerPublicKey: legacyED25519SenderAccount.publicKey,
           transaction: rawTxn,
@@ -516,13 +529,14 @@ describe("transaction simulation", () => {
       test("with multisig payload", async () => {
         const rawTxn = await aptos.generateTransaction({
           sender: legacyED25519SenderAccount.accountAddress.toString(),
-          feePayerAddress: feePayerAccount.accountAddress.toString(),
+          hasFeePayer: true,
           data: {
             multisigAddress: secondarySignerAccount.accountAddress,
             function: `${contractPublisherAccount.accountAddress.toString()}::transfer::transfer`,
             functionArguments: [new U64(1), recieverAccounts[0].accountAddress],
           },
         });
+        rawTxn.feePayerAddress = feePayerAccount.accountAddress;
 
         const [response] = await aptos.simulateTransaction({
           signerPublicKey: legacyED25519SenderAccount.publicKey,
@@ -535,7 +549,7 @@ describe("transaction simulation", () => {
         const rawTxn = await aptos.generateTransaction({
           sender: legacyED25519SenderAccount.accountAddress.toString(),
           secondarySignerAddresses: [secondarySignerAccount.accountAddress.toString()],
-          feePayerAddress: feePayerAccount.accountAddress.toString(),
+          hasFeePayer: true,
           data: {
             function: `${contractPublisherAccount.accountAddress.toString()}::transfer::two_by_two`,
             functionArguments: [
@@ -547,6 +561,7 @@ describe("transaction simulation", () => {
             ],
           },
         });
+        rawTxn.feePayerAddress = feePayerAccount.accountAddress;
 
         const [response] = await aptos.simulateTransaction({
           signerPublicKey: legacyED25519SenderAccount.publicKey,